Add layout links and tests
The important thing here for me to remember is that assert_select
whatever html element and that I can use different types of selectors.

The one used in this case (the `?` in `href=?`) is a substitute that is
replaced by the second parameter.

Search terms:
`assert_select edge api`
`html selectors edge api`